requirements.txt
# --- CORE SYSTEM ---
PyQt5==5.15.10
cryptography==42.0.8
argon2-cffi==23.1.0
pysqlcipher3==1.3.0

# --- CLOUD & SYNC ---
supabase==2.3.0
postgrest==0.16.2
gotrue==2.4.0
requests==2.31.0
httpx==0.26.0

# --- SECURITY & UTILS ---
pyotp==2.9.0
python-dotenv==1.0.0
python-dateutil==2.8.2
SpeechRecognition==3.10.0
PyAudio==0.2.14

# --- DEVELOPMENT & TESTING ---
pytest==7.4.3
pytest-asyncio==0.23.2
black==23.12.1
flake8==6.1.0
